Renames some components
