+++ client/src/const/theme.js
@@ -1,0 +1,17 @@
+const theme = {
+  bg: {
+    default: '#FFFFFF',
+    wash: '#FBFCFE',
+    reverse: '#282560'
+  },
+  text: {
+    default: '#111111',
+    secondary: '#646464',
+    placeholder: '#DCDBDB'
+  },
+  warn: {
+    default: '#E22F2F'
+  }
+}
+
+export default theme
